fix: :bug: fixed favicon
